Move `engine/envelope` package to `testkit/envelope`.
