+++ engine/engine.go
@@ -8,7 +8,7 @@ import (
 	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/cosyne"
 	"github.com/dogmatiq/dogma"
-	"github.com/dogmatiq/testkit/engine/envelope"
+	"github.com/dogmatiq/testkit/envelope"
 	"github.com/dogmatiq/testkit/fact"
 	"go.uber.org/multierr"
 )
